add movieId logic with saving data
